* pullup r19538.
	* transcode.c (rb_transcode_convertible): follow changes.

	* win32/win32.c (rb_w32_parse_cmdline, rb_w32_write_console): ditto.



git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/win32-unicode-test@19539 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ test/win32ole/test_win32ole_event.rb
@@ -6,6 +6,8 @@ require 'test/unit'
 
 if defined?(WIN32OLE_EVENT)
   class TestWIN32OLE_EVENT < Test::Unit::TestCase
+    module IE
+    end
     def create_temp_html
       fso = WIN32OLE.new('Scripting.FileSystemObject')
       dummy_file = fso.GetTempName + ".html"
@@ -17,9 +19,25 @@ if defined?(WIN32OLE_EVENT)
       dummy_path
     end
 
+    def message_loop
+      WIN32OLE_EVENT.message_loop
+      sleep 0.1
+    end
+
+    def wait_ie
+      while @ie.readyState != IE::READYSTATE_COMPLETE
+        message_loop
+      end
+    end
+
     def setup
+      WIN32OLE_EVENT.message_loop
       @ie = WIN32OLE.new("InternetExplorer.Application")
+      if !defined?(IE::READYSTATE_COMPLETE)
+        WIN32OLE.const_load(@ie, IE)
+      end
       @ie.visible = true
+      message_loop
       @event = ""
       @event2 = ""
       @event3 = ""
@@ -30,6 +48,12 @@ if defined?(WIN32OLE_EVENT)
       @event += event
     end
 
+    def test_s_new
+      assert_raise(TypeError) {
+        ev = WIN32OLE_EVENT.new("A")
+      }
+    end
+
     def test_s_new_without_itf
       ev = WIN32OLE_EVENT.new(@ie)
       ev.on_event {|*args| default_handler(*args)}
@@ -37,8 +61,7 @@ if defined?(WIN32OLE_EVENT)
       while @ie.busy
         WIN32OLE_EVENT.new(@ie)
         GC.start  
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
+        message_loop
       end
       assert_match(/BeforeNavigate/, @event)
       assert_match(/NavigateComplete/, @event)
@@ -48,13 +71,19 @@ if defined?(WIN32OLE_EVENT)
       ev = WIN32OLE_EVENT.new(@ie, 'DWebBrowserEvents')
       ev.on_event {|*args| default_handler(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.new(@ie, 'DWebBrowserEvents') 
-        GC.start  
-        sleep 0.1
-      end
+      wait_ie
       assert_match(/BeforeNavigate/, @event)
       assert_match(/NavigateComplete/, @event)
+    end
+
+    def test_on_event_symbol
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event(:BeforeNavigate2) {|*args|
+        handler1
+      }
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal("handler1", @event2)
     end
 
     def test_on_event2
@@ -62,9 +91,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event('BeforeNavigate') {|*args| handler1}
       ev.on_event('BeforeNavigate') {|*args| handler2}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-      end
+      wait_ie
       assert_equal("handler2", @event2)
     end
 
@@ -73,9 +100,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event {|*args| handler1}
       ev.on_event {|*args| handler2}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-      end
+      wait_ie
       assert_equal("handler2", @event2)
     end
 
@@ -85,9 +110,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event{|*args| handler2}
       ev.on_event('NavigateComplete'){|*args| handler3(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-      end
+      wait_ie
       assert(@event3!="")
       assert("handler2", @event2)
     end
@@ -97,9 +120,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event {|*args| default_handler(*args)}
       ev.on_event('NavigateComplete'){|*args| handler3(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-      end
+      wait_ie
       assert_match(/BeforeNavigate/, @event)
       assert(/NavigateComplete/ !~ @event)
       assert(@event!="")
@@ -109,16 +130,12 @@ if defined?(WIN32OLE_EVENT)
       ev = WIN32OLE_EVENT.new(@ie, 'DWebBrowserEvents')
       ev.on_event {|*args| default_handler(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-      end
+      wait_ie
       assert_match(/BeforeNavigate/, @event)
       ev.unadvise
       @event = ""
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-      end
+      wait_ie
       assert_equal("", @event);
       assert_raise(WIN32OLERuntimeError) {
         ev.on_event {|*args| default_handler(*args)}
@@ -135,6 +152,102 @@ if defined?(WIN32OLE_EVENT)
       }
     end
 
+    def test_on_event_with_outargs
+      ev = WIN32OLE_EVENT.new(@ie)
+      # ev.on_event_with_outargs('BeforeNavigate'){|*args|
+      #  args.last[5] = true # Cancel = true
+      # }
+      ev.on_event_with_outargs('BeforeNavigate2'){|*args|
+        args.last[6] = true # Cancel = true
+      }
+      bl = @ie.locationURL
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal(bl, @ie.locationURL)
+    end
+
+    def test_on_event_hash_return
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){|*args|
+        {:return => 1, :Cancel => true}
+      }
+      bl = @ie.locationURL
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal(bl, @ie.locationURL)
+    end
+
+    def test_on_event_hash_return2
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){|*args|
+        {:Cancel => true}
+      }
+      bl = @ie.locationURL
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal(bl, @ie.locationURL)
+    end
+
+    def test_on_event_hash_return3
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){|*args|
+        {'Cancel' => true}
+      }
+      bl = @ie.locationURL
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal(bl, @ie.locationURL)
+    end
+    
+    def test_on_event_hash_return4
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){|*args|
+        {'return' => 2, 'Cancel' => true}
+      }
+      bl = @ie.locationURL
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal(bl, @ie.locationURL)
+    end
+
+    def test_on_event_hash_return5
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){|*args|
+        {6 => true}
+      }
+      bl = @ie.locationURL
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal(bl, @ie.locationURL)
+    end
+
+    def test_off_event
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event{handler1}
+      ev.off_event
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal("", @event2)
+    end
+
+    def test_off_event_arg
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){handler1}
+      ev.off_event('BeforeNavigate2')
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal("", @event2)
+    end
+
+    def test_off_event_sym_arg
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){handler1}
+      ev.off_event(:BeforeNavigate2)
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal("", @event2)
+    end
+
     def handler1
       @event2 = "handler1"
     end
@@ -149,9 +262,72 @@ if defined?(WIN32OLE_EVENT)
 
     def teardown
       @ie.quit
+      message_loop
       @ie = nil
-      File.unlink(@f)
+      i = 0
+      begin 
+        i += 1
+        File.unlink(@f) if i < 10
+      rescue Errno::EACCES
+        message_loop
+        retry
+      end
+      message_loop
       GC.start
+      message_loop
     end
+
+    class Handler1
+      attr_reader :val1, :val2, :val3, :val4
+      def initialize
+        @val1 = nil
+        @val2 = nil
+        @val3 = nil
+        @val4 = nil
+      end
+      def onStatusTextChange(t)
+        @val1 = t
+      end
+      def onProgressChange(p, pmax)
+        @val2 = p
+        @val3 = pmax
+      end
+      def onPropertyChange(p)
+        @val4 = p
+      end
+    end
+
+    class Handler2
+      attr_reader :ev
+      def initialize
+        @ev = ""
+      end
+      def method_missing(ev, *arg)
+        @ev += ev
+      end
+    end
+
+    def test_handler1
+      ev = WIN32OLE_EVENT.new(@ie)
+      h1 = Handler1.new
+      ev.handler = h1
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert(h1.val1)
+      assert_equal(h1.val1, ev.handler.val1)
+      assert(h1.val2)
+      assert(h1.val3)
+      assert(h1.val4)
+    end
+
+    def test_handler2
+      ev = WIN32OLE_EVENT.new(@ie)
+      h2 = Handler2.new
+      ev.handler = h2
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert(h2.ev != "")
+    end
+
   end
 end
